manage unit of measurement
'-u' option can be set to adapt display according to wanted size unit
(KB, MB, GB or TB).
KB is set by default if option is not explicitly defined.
When KB or MB is set, size is displayed as unsigned integer.
When GB or TB is set, size is displayed as float with 2 digits of
precision.